adds Vpd setpoint for output power (#4)
* adds Vpd setpoint for output power

* fix serial baud switching

+++ src/src/gd32f1x0/pwm.c
@@ -123,7 +123,7 @@ gpio_pwm_t pwm_init(uint32_t pin)
     return (gpio_pwm_t){.tim = timer, .ch = channel};
 }
 
-void pwm_out_write(gpio_pwm_t pwm, int val)
+void pwm_out_write(gpio_pwm_t pwm, uint16_t val)
 {
     if (pwm.tim) {
         // Register value is target-1
